refactor(node): move api test_helpers and api_tests to own files
